Usage of ConfigOptionProvider instead of ConfigDescriptionProvider in order to provide options
