+++ addons/binding/org.openhab.binding.amazondashbutton/ESH-INF/thing/thing-types.xml
@@ -12,6 +12,10 @@
 			<channel id="press" typeId="trigger-channel" />
 		</channels>
 		<config-description>
+			<parameter name="pcapNetworkInterfaceName" type="text">
+				<label>@text/dashButtonNetworkInterfaceLabel</label>
+				<description>@text/dashButtonNetworkInterfaceDescription</description>
+			</parameter>
 			<parameter name="macAddress" type="text" pattern="([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})">
 				<label>@text/dashButtonMacAddressLabel</label>
 				<description>@text/dashButtonMacAddressDescription</description>
