+++ addons/binding/org.openhab.binding.amazondashbutton/META-INF/MANIFEST.MF
@@ -4,13 +4,14 @@ Bundle-Name: AmazonDashButton Binding
 Bundle-SymbolicName: org.openhab.binding.amazondashbutton;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.0.0.qualifier
-Bundle-Activator: org.openhab.binding.amazondashbutton.internal.Activator
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ClassPath: .,
  lib/jna-4.2.1.jar,
  lib/pcap4j-core-1.6.6.jar,
  lib/pcap4j-packetfactory-static-1.6.6.jar
 Import-Package: 
+ com.google.common.base,
+ com.google.common.collect,
  org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.config.core.i18n,
  org.eclipse.smarthome.config.discovery,
