+++ addons/binding/org.openhab.binding.amazondashbutton/OSGI-INF/AmazonDashButtonConfigOptionProvider.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2014-2016 by the respective copyright holders.
+
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.binding.amazondashbutton.internal.config.AmazonDashButtonConfigOptionProvider">
+
+   <implementation class="org.openhab.binding.amazondashbutton.internal.config.AmazonDashButtonConfigOptionProvider"/>
+
+   <service>
+      <provide interface="org.eclipse.smarthome.config.core.ConfigOptionProvider"/>
+   </service>
+
+</scr:component>
